Renamed serializer files for naming consistency
